requirements.txt
pillow
numpy
matplotlib
torch
torchvision